+++ src/main/java/org/oagi/srt/gateway/http/api/bie_management/data/BieCreateRequest.java
@@ -3,13 +3,14 @@ package org.oagi.srt.gateway.http.api.bie_management.data;
 import lombok.Data;
 import org.jooq.types.ULong;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Data
 public class BieCreateRequest {
 
-    private long asccpManifestId;
-    private List<Long> bizCtxIds;
+    private BigInteger asccpManifestId;
+    private List<BigInteger> bizCtxIds;
 
     public ULong asccpManifestId() {
         return ULong.valueOf(asccpManifestId);
