+++ src/main/java/org/oagi/srt/gateway/http/api/bie_management/data/DeleteBieListRequest.java
@@ -2,12 +2,13 @@ package org.oagi.srt.gateway.http.api.bie_management.data;
 
 import lombok.Data;
 
+import java.math.BigInteger;
 import java.util.Collections;
 import java.util.List;
 
 @Data
 public class DeleteBieListRequest {
 
-    private List<Long> topLevelAbieIds = Collections.emptyList();
+    private List<BigInteger> topLevelAbieIds = Collections.emptyList();
 
 }
