+++ src/main/java/org/oagi/srt/gateway/http/api/bie_management/data/bie_edit/BccForBie.java
@@ -4,11 +4,13 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import org.oagi.srt.gateway.http.api.common.data.TrackableImpl;
 
+import java.math.BigInteger;
+
 @Data
 @EqualsAndHashCode(callSuper = true)
 public class BccForBie extends TrackableImpl {
 
-    private long bccId;
+    private BigInteger bccId;
     private String guid;
     private String cardinalityMin;
     private String cardinalityMax;
@@ -19,7 +21,7 @@ public class BccForBie extends TrackableImpl {
     private String fromAccId;
 
     @Override
-    public long getId() {
+    public BigInteger getId() {
         return bccId;
     }
