+++ src/main/java/org/oagi/srt/gateway/http/api/bie_management/data/bie_edit/BieEditAgencyIdList.java
@@ -2,10 +2,12 @@ package org.oagi.srt.gateway.http.api.bie_management.data.bie_edit;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class BieEditAgencyIdList {
 
-    private long agencyIdListId;
+    private BigInteger agencyIdListId;
     private boolean isDefault;
     private String agencyIdListName;
