+++ src/main/java/org/oagi/srt/gateway/http/api/bie_management/data/bie_edit/BieEditBbiep.java
@@ -2,10 +2,12 @@ package org.oagi.srt.gateway.http.api.bie_management.data.bie_edit;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class BieEditBbiep {
 
-    private long bbiepId;
-    private long basedBccpId;
+    private BigInteger bbiepId;
+    private BigInteger basedBccpId;
 
 }
