+++ src/main/java/org/oagi/srt/gateway/http/api/bie_management/data/bie_edit/BieEditBcc.java
@@ -6,15 +6,17 @@ import org.oagi.srt.data.Cardinality;
 import org.oagi.srt.data.SeqKeySupportable;
 import org.oagi.srt.gateway.http.api.common.data.TrackableImpl;
 
+import java.math.BigInteger;
+
 @Data
 @EqualsAndHashCode(callSuper = true)
 public class BieEditBcc extends TrackableImpl implements SeqKeySupportable, Cardinality {
 
-    private long bccId;
-    private long bccManifestId;
+    private BigInteger bccId;
+    private BigInteger bccManifestId;
     private String guid;
-    private long fromAccManifestId;
-    private long toBccpManifestId;
+    private BigInteger fromAccManifestId;
+    private BigInteger toBccpManifestId;
     private int seqKey;
     private int entityType;
 
@@ -26,7 +28,7 @@ public class BieEditBcc extends TrackableImpl implements SeqKeySupportable, Card
     }
 
     @Override
-    public long getId() {
+    public BigInteger getId() {
         return bccId;
     }
