+++ src/main/java/org/oagi/srt/gateway/http/api/bie_management/data/expression/GenerateExpressionOption.java
@@ -2,6 +2,8 @@ package org.oagi.srt.gateway.http.api.bie_management.data.expression;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class GenerateExpressionOption {
 
@@ -19,20 +21,20 @@ public class GenerateExpressionOption {
 
     private boolean arrayForJsonExpression;
     private boolean includeMetaHeaderForJson;
-    private Long metaHeaderTopLevelAbieId;
+    private BigInteger metaHeaderTopLevelAbieId;
     private boolean includePaginationResponseForJson;
-    private Long paginationResponseTopLevelAbieId;
+    private BigInteger paginationResponseTopLevelAbieId;
 
     private String openAPIExpressionFormat;
     private boolean openAPI30GetTemplate;
     private boolean arrayForJsonExpressionForOpenAPI30GetTemplate;
     private boolean includeMetaHeaderForJsonForOpenAPI30GetTemplate;
-    private Long metaHeaderTopLevelAbieIdForOpenAPI30GetTemplate;
+    private BigInteger metaHeaderTopLevelAbieIdForOpenAPI30GetTemplate;
     private boolean includePaginationResponseForJsonForOpenAPI30GetTemplate;
-    private Long paginationResponseTopLevelAbieIdForOpenAPI30GetTemplate;
+    private BigInteger paginationResponseTopLevelAbieIdForOpenAPI30GetTemplate;
 
     private boolean openAPI30PostTemplate;
     private boolean arrayForJsonExpressionForOpenAPI30PostTemplate;
     private boolean includeMetaHeaderForJsonForOpenAPI30PostTemplate;
-    private Long metaHeaderTopLevelAbieIdForOpenAPI30PostTemplate;
+    private BigInteger metaHeaderTopLevelAbieIdForOpenAPI30PostTemplate;
 }
