+++ src/main/java/org/oagi/srt/gateway/http/api/bie_management/service/edit_tree/BieEditTreeController.java
@@ -5,11 +5,12 @@ import org.oagi.srt.gateway.http.api.bie_management.data.bie_edit.BieEditNode;
 import org.oagi.srt.gateway.http.api.bie_management.data.bie_edit.tree.BieEditAbieNode;
 import org.oagi.srt.gateway.http.api.bie_management.data.bie_edit.tree.BieEditNodeDetail;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface BieEditTreeController {
 
-    BieEditAbieNode getRootNode(long topLevelAbieId);
+    BieEditAbieNode getRootNode(BigInteger topLevelAbieId);
 
     List<BieEditNode> getDescendants(BieEditNode node, boolean hideUnused);
