+++ src/main/java/org/oagi/srt/gateway/http/api/common/data/TrackableImpl.java
@@ -3,13 +3,15 @@ package org.oagi.srt.gateway.http.api.common.data;
 import lombok.Data;
 import org.oagi.srt.data.Trackable;
 
+import java.math.BigInteger;
+
 @Data
 public abstract class TrackableImpl implements Trackable {
 
-    private long releaseId;
+    private BigInteger releaseId;
     private String releaseNum;
 
-    private long revisionId;
+    private BigInteger revisionId;
     private int revisionNum;
     private int revisionTrackingNum;
